+++ src/components/Unit/Store/StoreCreateAndUpdate/StoreCreateAndUpdate.container.tsx
@@ -1,0 +1,5 @@
+import StoreCreateAndUpdateUI from "./StoreCreateAndUpdate.presenter";
+
+export default function StoreCreateAndUpdate(): JSX.Element {
+  return <StoreCreateAndUpdateUI />;
+}
