+++ src/components/Unit/Store/StoreCreateAndUpdate/StoreCreateAndUpdate.presenter.tsx
@@ -1,0 +1,3 @@
+export default function StoreCreateAndUpdateUI(): JSX.Element {
+  return <></>;
+}
