+++ src/components/Unit/Store/StoreList/StoreList.container.tsx
@@ -1,5 +1,11 @@
+import { useRouter } from "next/router";
 import StoreListUI from "./StoreList.presenter";
 
 export default function StoreList(): JSX.Element {
-  return <StoreListUI />;
+  const router = useRouter();
+  const onClickStoreItem = (): void => {
+    router.push("/store/create");
+  };
+
+  return <StoreListUI onClickStoreItem={onClickStoreItem} />;
 }
