+++ src/pages/store/[id]/index.tsx
@@ -1,1 +1,1 @@
-
+// export default function
