+++ src/types/store/index.ts
@@ -1,0 +1,8 @@
+export interface StoreType {
+  id?: string;
+  name: string;
+  description: string;
+  create_date?: string;
+  owner?: string;
+  address?: string;
+}
